*: add initial implementation
This is a pretty simple server implementation that accepts tarballs
POSTed to it, runs `syft` externally, and returns the generated
SPDX-formatted SBOM. By default, `syft` is expected to be in the root
directory (i.e. `sbom-server` is run from within a container whose
base image is hub.docker.io/anchore/syft) but to aid development, the
official container image can be run using docker instead.