git: add mechanism for retrying a commit message
Previously, if the author attempted to make a commit with an improper
log message, the commit was aborted and they had to try again. This is
a frustrating experience, especially if they are unaware that their
previous attempt was saved in `.git/COMMIT_EDITMSG`. This tries to
make that process a little easier by making a backup of the message
and then restoring it when the next attempt at a commit is made.