in-toto: fix statement subjects
The subject of the SPDX is the source code; not the SPDX. Inversely,
the subject of the SCAI is the SPDX; not the source code. The artifact
name has also been updated to include the `tar` extension, since it's
a tar archive of the container image, rather than its native format.
This should make it more clear that the associated digest is that of
the tarball, rather than the container image digest.