server: derive artifact name from container image
This now uses the first image reference found in both docker and OCI
container images as the name of the artifact. It should make it a
little easier to associate an SBOM to an artifact, though it doesn't
change the validation process since the artifact name is just a hint.