in-toto: add SLSA provenance
This document explains the parameters of the SLSA Build (generating
the SBOM). While it doesn't technically provide any new information,
it does make it much easier for downstream consumers to understand
what went into the generation of an SBOM (i.e. the version numbers and
the arguments to the server).