nsm: add mock implementation
This makes it easier to test the server locally and it will probably
make testing easier too. The nsm config argument is skipped in the
Provenance document because it will always be "real" - if it's fake,
the SCAI attribute evidence won't validate. The COSE_Sign1 CBOR object
is being manually constructed to work around an upstream bug (if the
protected attribute is an empty map, it gets serialized as an empty
bstr instead of a bstr-encoded empty map). Once that's fixed, the
implementation can be changed to:

```rust
Ok(serde_cbor::to_vec(&Sign1 {
    protected: Map(BTreeMap::from([(HeaderParameter::KID, 0)])),
    unprotected: Map::new(),
    payload: attestation.to_binary(),
    signature: Vec::new(),
})
```