in_toto: preserve the SPDX's byte sequence
Attempting to validate a response from the server was resulting in a
digest mismatch between the SPDX and the SCAI referencing it in its
subject. The SPDX that comes back from syft is pretty printed, so
there are newlines and spaces scattered throughout. When serde_json
serializes this object, it removes all of the formatting and outputs a
minimal representation. This causes a digest mismatch, since the
digest is calculated from the original JSON before it is formatted.

The `serde_json::json!` macro apparently is responsible for this
formatting. It formats all of its parameters (I think via
`serde_json::to_string()`), even if the type is RawValue. This doesn't
happen when the RawValue is an attribute of a struct, so the
implementation was updated to use that method instead.

This is an (ill-advised) attempt at stabilizing the format of the
predicate - just treating it as a string, even though it is an
object. The alternative is to reference the in-toto Envelope's
payload, instead of the inner SPDX, since its base64-encoded and
therefore well-defined. I decided against that route because I
couldn't think of a decent way to reference the Envelope payload from
the SCAI Statement's subject.