in-toto: properly encode SPDX predicate
The SPDX is JSON-encoded and should be included as an object rather
than a string. The server was returning this:

    "predicate":"{\n \"spdxVersion\": \"SPDX-2.3\",...

when it should have returned:

    "predicate":{"spdxVersion":"SPDX-2.3",...

A side effect of this representation is that the predicate is no
longer fully defined - there is no requirement on the byte sequence of
the encoded JSON. This will need to be addressed before verification
can be done reliably.