client: calculate the digest of the artifact
This calculation is slightly tricky to do. The request body is a
stream which is consumed by reqwest, but the stream also needs to
incrementally update a running digest as it's polled. Reading the
digest after the stream has hit the end requires having a reference to
that original stream. I tried implementing this as a tokio watch
channel, but it's not possible (with the current crates in use) to
repeatedly finalize the digest - it's consumed in the process. The
Arc-wrapped Mutex provides the escape hatch instead, though it's a bit
of a kludge.