client: verify server hardening
The Provenance Document includes the arguments passed to the
server. This verifies that they are acceptable values. This could also
be verified by looking at the builder ID, but it becomes slightly
tricky to do when the client is a different version from the server -
the version is encoded in the builder ID, so it wouldn't be a simple
comparison.