github: add release workflow
This defines the release process for the project. It builds the client
and server, builds the enclave image, and publishes the artifacts to
AWS ECR and GitHub. The server and client are built for both x86_64 and
aarch64, but the enclave image is x86_64 only.

This process is complicated by the fact that the client needs to know
the PCR0 of the enclave image. To facilitate this, the server has to
be built first; followed by the enclave image, taking note of the
PCR0; and then the client is built with the newly calculated PCR0s.
After the release process finishes, these digests will then need to be
committed to the repo so future builds continue to trust them.

The biggest complication to this process stems from the fact that
Ubuntu doesn't currently publish binary assets for musl-dev:arm64 and
GitHub workers can't be customized to run a different OS. As a result,
part of the build needs to happen in a Debian container, since _they_
publish the musl-dev:arm64 assets. Unfortunately, the process of
building an enclave image requires docker, which means that it can't
run within this containerized environment. Pulling this all together,
the build starts in a containerized environment, transfers the assets
to a native worker, and then transfers the assets again to another
containerized environment.