github: build clients for MacOS
I couldn't figure out how to get the re-usable "build" workflow to use
the matrix strategy, so I moved (and duplicated) the matrix in the
"release" workflow. Jobs can't be conditional on matrix elements
because the condition is evaluated before the matrix [^1] and it
doesn't appear to be possible to conditionally build the matrix based
on an input value.

[^1]: https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idif